hw13 - separation AppConfig
