+++ hw13-L25-DI/src/main/java/ru/otus/config/AppConfig1.java
@@ -1,0 +1,24 @@
+package ru.otus.config;
+
+import ru.otus.appcontainer.api.AppComponent;
+import ru.otus.appcontainer.api.AppComponentsContainerConfig;
+import ru.otus.services.*;
+
+@AppComponentsContainerConfig(order = 1)
+public class AppConfig1 {
+
+    @AppComponent(order = 0, name = "equationPreparer")
+    public EquationPreparer equationPreparer() {
+        return new EquationPreparerImpl();
+    }
+
+    @AppComponent(order = 1, name = "ioService")
+    public IOService ioService() {
+        return new IOServiceStreams(System.out, System.in);
+    }
+
+    @AppComponent(order = 2, name = "playerService")
+    public PlayerService playerService(IOService ioService) {
+        return new PlayerServiceImpl(ioService);
+    }
+}
