+++ hw13-L25-DI/src/main/java/ru/otus/config/AppConfig2.java
@@ -1,0 +1,15 @@
+package ru.otus.config;
+
+import ru.otus.appcontainer.api.AppComponent;
+import ru.otus.appcontainer.api.AppComponentsContainerConfig;
+import ru.otus.services.*;
+
+@AppComponentsContainerConfig(order = 2)
+public class AppConfig2 {
+
+    @AppComponent(order = 0, name = "gameProcessor")
+    public GameProcessor gameProcessor(
+            IOService ioService, PlayerService playerService, EquationPreparer equationPreparer) {
+        return new GameProcessorImpl(ioService, equationPreparer, playerService);
+    }
+}
