+++ hw13-L25-DI/src/test/java/ru/otus/AppTest.java
@@ -7,7 +7,8 @@ import org.junit.jupiter.params.provider.CsvSource;
 import ru.otus.appcontainer.AppComponentsContainerImpl;
 import ru.otus.appcontainer.api.AppComponent;
 import ru.otus.appcontainer.api.AppComponentsContainerConfig;
-import ru.otus.config.AppConfig;
+import ru.otus.config.AppConfig1;
+import ru.otus.config.AppConfig2;
 import ru.otus.services.*;
 
 import java.io.PrintStream;
@@ -40,7 +41,7 @@ class AppTest {
             })
     void shouldExtractFromContextCorrectComponentWithNotNullFields(String classNameOrBeanId, Class<?> rootClass)
             throws Exception {
-        var ctx = new AppComponentsContainerImpl(AppConfig.class);
+        var ctx = new AppComponentsContainerImpl(AppConfig1.class, AppConfig2.class);
 
         assertThat(classNameOrBeanId).isNotEmpty();
         Object component;
